+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Unreleased
+----------
+ 
+ - (Added) Priority can be set for transitions (using *low*, *high* or any integer in yaml). transitions
+   are selected according to their priorities (still following eventless and inner-first/source state semantics).
+
+
 1.2.2 (2018-06-21)
 ------------------